+++ exporters/stdout/exporter.go
@@ -20,7 +20,6 @@ import (
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/metric/global"
 	"go.opentelemetry.io/otel/sdk/export/metric"
-	exporttrace "go.opentelemetry.io/otel/sdk/export/trace"
 	controller "go.opentelemetry.io/otel/sdk/metric/controller/basic"
 	processor "go.opentelemetry.io/otel/sdk/metric/processor/basic"
 	"go.opentelemetry.io/otel/sdk/metric/selector/simple"
@@ -34,8 +33,8 @@ type Exporter struct {
 }
 
 var (
-	_ metric.Exporter          = &Exporter{}
-	_ exporttrace.SpanExporter = &Exporter{}
+	_ metric.Exporter       = &Exporter{}
+	_ sdktrace.SpanExporter = &Exporter{}
 )
 
 // NewExporter creates an Exporter with the passed options.